HS_Script/HS_Script/requirements.txt
# Hearthstone Bot Dependencies
# Install: pip install -r requirements.txt

# Screen capture and automation
pillow>=10.0.0
opencv-python>=4.8.0
mss>=9.0.1

# Mouse/keyboard automation
pyautogui>=0.9.54
pynput>=1.7.6

# Log parsing
watchdog>=3.0.0

# UI overlay
PyQt6>=6.5.0

# Data
requests>=2.31.0
numpy>=1.24.0

# Tools
loguru>=0.7.0
pydantic>=2.0.0
